exception handling

+++ Exception_Handling/app1/src/H1.java
@@ -3,7 +3,7 @@ class H1
 	public static void main(String[] args) 
 	{
 		System.out.println("main begin");
-		int[] x = new int[9999999999];
+		long[] x = new long[9999999999];
 		System.out.println("main end");
 	}
 }
